+++ src/ZLinq/Linq/All.cs
@@ -50,7 +50,7 @@
             // The assembly output differs slightly when iterating through array and span.
             // According to microbenckmark results, iterating through span was faster for the logic passed to predicate.
             var span = (ReadOnlySpan<TSource>)array;
-            for (int i = 0; i < span.Length; i++)
+            for (int i = 0; (uint)i < (uint)span.Length; i++)
             {
                 if (!predicate(span[i]))
                 {
@@ -62,7 +62,7 @@
             [MethodImpl(MethodImplOptions.NoInlining)]
             static Boolean All(TSource[] array, Func<TSource, Boolean> predicate)
             {
-                for (int i = 0; i < array.Length; i++)
+                for (int i = 0; (uint)i < (uint)array.Length; i++)
                 {
                     if (!predicate(array[i]))
                     {
