+++ src/ZLinq/Linq/Max.cs
@@ -301,7 +301,7 @@ partial class ValueEnumerableExtensions
         else
         {
             var result = span[0];
-            for (int i = 1; i < span.Length; i++)
+            for (int i = 1; (uint)i < (uint)span.Length; i++)
             {
                 if (span[i] > result)
                 {
