+++ src/ZLinq/Linq/Min.cs
@@ -297,7 +297,7 @@ partial class ValueEnumerableExtensions
         else
         {
             var result = span[0];
-            for (int i = 1; i < span.Length; i++)
+            for (int i = 1; (uint)i < (uint)span.Length; i++)
             {
                 if (span[i] < result)
                 {
